Add functions to convert categorical features into one-hots and perform nested cross-validation.
Currently the nested cross validation does not use the categorical one-hot function. That needs to be added. The difficulty is processing the data appropriately to account for different category values appearing in the training vs test set. The easiest course of action is to bite the bullet and assume no new categories can appear when we get a new sample outside this dataset. Alternatively, we could add preprocessing such that unseen values are labeled 'other' or something.